fix(shared): keep auth imports free of compile-time UI macros
