+++ apps/desktop/src/renderer/routes/_authenticated/settings/members/components/MembersSettings/utils/organizationRoleName/index.ts
@@ -1,0 +1,1 @@
+export { organizationRoleName } from "./organizationRoleName";
