+++ packages/shared/src/auth/roles/roles.ts
@@ -1,6 +1,3 @@
-import { msg } from "@lingui/core/macro";
-import { i18n } from "../../i18n";
-
 // Role hierarchy from lowest to highest permission
 export const ROLE_HIERARCHY = ["member", "admin", "owner"] as const;
 
@@ -14,22 +11,6 @@ export const ORGANIZATION_ROLES: Record<
 	admin: { id: "admin", name: "Admin" },
 	owner: { id: "owner", name: "Owner" },
 };
-
-/**
- * The role's display name in the active locale. `ORGANIZATION_ROLES[x].name`
- * stays plain English: it is stable data (logs, server payloads), so display
- * code renders this instead.
- */
-export function organizationRoleName(role: OrganizationRole): string {
-	switch (role) {
-		case "owner":
-			return i18n._(msg({ message: "Owner" }));
-		case "admin":
-			return i18n._(msg({ message: "Admin" }));
-		case "member":
-			return i18n._(msg({ message: "Member" }));
-	}
-}
 
 export function getRoleLevel(role: OrganizationRole): number {
 	return ROLE_HIERARCHY.indexOf(role);
